% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplsr-package.R
\docType{package}
\name{aplsr-package}
\alias{aplsr}
\alias{aplsr-package}
\title{aplsr: Shared R Helpers for the AP-LS Website}
\description{
Reusable helper functions used to build the American Psychology-Law Society (AP-LS) Quarto website, including award "past recipients" tables, small HTML/UI components, and utilities. Extracted from the site's assets/apls_functions.R so pages can call library(aplsr) instead of source()-ing a script.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/APLS-org/apls-r}
  \item Report bugs at \url{https://github.com/APLS-org/apls-r/issues}
}

}
\author{
\strong{Maintainer}: AP-LS Website Editor \email{office@ap-ls.org} (Update to the current maintainer)

Authors:
\itemize{
  \item American Psychology-Law Society [copyright holder]
}

}
\keyword{internal}
